Move Validator Base classes into a dedicated directory.
