+++ trpcultivate_phenotypes/src/TripalCultivateValidator/TripalCultivatePhenotypesValidatorInterface.php
@@ -1,13 +1,6 @@
 <?php
 
-/**
- * @file
- * Contains \Drupal\trpcultivate_phenotypes\Interface\TripalCultivatePhenotypesValidatorInterface.
- * 
- * @see Plugin manager in src\TripalCultivatePhenotypesValidatorManager.php
- */
-
-namespace Drupal\trpcultivate_phenotypes;
+namespace Drupal\trpcultivate_phenotypes\TripalCultivateValidator;
 
 use Drupal\Component\Plugin\PluginInspectionInterface;
 
@@ -19,20 +12,20 @@ interface TripalCultivatePhenotypesValidatorInterface extends PluginInspectionIn
    * Return the name of the validator.
    *
    * @return string.
-   */    
+   */
   public function getValidatorName();
 
   /**
    * Return the scope of the validator.
    *
    * @return string.
-   */    
+   */
   public function getValidatorScope();
 
   /**
    * Load data file import assets Project title, Genus and Data File Id
    * as entered in the Importer form.
-   * 
+   *
    * @param $project
    *   String, Project name/title - chado.project: name.
    * @param $genus
@@ -45,7 +38,7 @@ interface TripalCultivatePhenotypesValidatorInterface extends PluginInspectionIn
    *   Boolean, skip flag when set to true will skip the validation
    *   logic and set the validator as upcoming/todo.
    *   Default: false - execute validation process.
-   * 
+   *
    * @return void.
    */
   public function loadAssets($project, $genus, $file_id, $headers, $skip);
