+++ trpcultivate_phenotypes/src/TripalCultivateValidator/TripalCultivatePhenotypesValidatorManager.php
@@ -1,11 +1,6 @@
 <?php
 
-/**
- * @file
- * Contains Tripal Cultivate Phenotypes Validator Plugin Manager.
- */
-
-namespace Drupal\trpcultivate_phenotypes;
+namespace Drupal\trpcultivate_phenotypes\TripalCultivateValidator;
 
 use Drupal\Core\Plugin\DefaultPluginManager;
 use Drupal\Core\Cache\CacheBackendInterface;
@@ -17,7 +12,7 @@ use Drupal\Core\Extension\ModuleHandlerInterface;
 class TripalCultivatePhenotypesValidatorManager extends DefaultPluginManager {
   /**
    *  Constructs Validator Plugin Manager.
-   * 
+   *
    *  @param \Traversable $namespaces
    *    An object that implements \Traversable which contains the root paths
    *    keyed by the corresponding namespace to look for plugin implementations.
@@ -28,10 +23,10 @@ class TripalCultivatePhenotypesValidatorManager extends DefaultPluginManager {
    */
   public function __construct(\Traversable $namespaces, CacheBackendInterface $cache_backend, ModuleHandlerInterface $module_handler) {
     parent::__construct(
-      'Plugin/Validators', 
-      $namespaces, 
-      $module_handler, 
-      'Drupal\trpcultivate_phenotypes\TripalCultivatePhenotypesValidatorInterface', 
+      'Plugin/Validators',
+      $namespaces,
+      $module_handler,
+      'Drupal\trpcultivate_phenotypes\TripalCultivatePhenotypesValidatorInterface',
       'Drupal\trpcultivate_phenotypes\Annotation\TripalCultivatePhenotypesValidator'
     );
 
@@ -47,10 +42,10 @@ class TripalCultivatePhenotypesValidatorManager extends DefaultPluginManager {
 
   /**
    * Retrieve validator implementation with a specific scope.
-   * 
+   *
    * @param string $scope
    *   The validator_scope you are interested in.
-   * 
+   *
    * @return string
    *   The id of the validator with that scope based on it's annotation.
    */
@@ -59,7 +54,7 @@ class TripalCultivatePhenotypesValidatorManager extends DefaultPluginManager {
     $plugin_definitions = array_values($plugins);
 
     $plugin_key = array_search(
-      $scope, 
+      $scope,
       array_column($plugin_definitions, 'validator_scope')
     );
