+++ trpcultivate_phenotypes/trpcultivate_phenotypes.services.yml
@@ -17,10 +17,10 @@ services:
     arguments: ['@config.factory', '@current_user']
 
   plugin.manager.trpcultivate_validator:
-    class: 'Drupal\trpcultivate_phenotypes\TripalCultivatePhenotypesValidatorManager'
+    class: 'Drupal\trpcultivate_phenotypes\TripalCultivateValidator\TripalCultivatePhenotypesValidatorManager'
     parent: default_plugin_manager
     arguments: ['@tripal_chado.database']
 
   trpcultivate_phenotypes.traits:
     class: 'Drupal\trpcultivate_phenotypes\Service\TripalCultivatePhenotypesTraitsService'
-    arguments: ['@trpcultivate_phenotypes.genus_ontology', '@trpcultivate_phenotypes.terms', '@tripal_chado.database', '@tripal.logger']  
+    arguments: ['@trpcultivate_phenotypes.genus_ontology', '@trpcultivate_phenotypes.terms', '@tripal_chado.database', '@tripal.logger']
